Make sure users can only delete their own imorted records

+++ app/models/import_history.rb
@@ -128,13 +128,21 @@ class ImportHistory < ActiveRecord::Base
     end
 
     # Method to delete imported records
-    def delete_imported_records
-	# Get import_data for this record
-	id=ImportDatum.where(:import_history_id => self.id).where(:process_flag => false)
-	# Destroy records
-	if id.destroy_all
-	    return true
+    def delete_imported_records(user_id)
+	# Check user
+	if self.user_id == user_id
+	    # Get import_data for this record
+	    id=ImportDatum.where(:import_history_id => self.id).where(:process_flag => false)
+	    # Destroy records
+	    if id.destroy_all
+		return true
+	    else
+		return false
+	    end
 	else
+	    # Wrong user
+	    self.errors.add(:base,"Cannot delete other user's imported records.")
+	    # Return nil
 	    return false
 	end
     end
